+++ src/scfeatureprofiler/__init__.py
@@ -16,4 +16,4 @@ __all__ = [
     "find_marker_features",
 ]
 
-__version__ = "1.0.3"
+__version__ = "1.1.0"